requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
asyncer==0.0.12
attrs==25.4.0
bidict==0.23.1
certifi==2026.1.4
cffi==2.0.0
chainlit==2.9.5
charset-normalizer==3.4.4
chevron==0.14.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.0
filelock==3.20.3
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.2
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.25.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.1
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-agno==0.50.1
opentelemetry-instrumentation-alephalpha==0.50.1
opentelemetry-instrumentation-anthropic==0.50.1
opentelemetry-instrumentation-bedrock==0.50.1
opentelemetry-instrumentation-chromadb==0.50.1
opentelemetry-instrumentation-cohere==0.50.1
opentelemetry-instrumentation-crewai==0.50.1
opentelemetry-instrumentation-google-generativeai==0.50.1
opentelemetry-instrumentation-groq==0.50.1
opentelemetry-instrumentation-haystack==0.50.1
opentelemetry-instrumentation-lancedb==0.50.1
opentelemetry-instrumentation-langchain==0.50.1
opentelemetry-instrumentation-llamaindex==0.50.1
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-marqo==0.50.1
opentelemetry-instrumentation-mcp==0.50.1
opentelemetry-instrumentation-milvus==0.50.1
opentelemetry-instrumentation-mistralai==0.50.1
opentelemetry-instrumentation-ollama==0.50.1
opentelemetry-instrumentation-openai==0.50.1
opentelemetry-instrumentation-openai-agents==0.50.1
opentelemetry-instrumentation-pinecone==0.50.1
opentelemetry-instrumentation-qdrant==0.50.1
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-replicate==0.50.1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sagemaker==0.50.1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-together==0.50.1
opentelemetry-instrumentation-transformers==0.50.1
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-vertexai==0.50.1
opentelemetry-instrumentation-watsonx==0.50.1
opentelemetry-instrumentation-weaviate==0.50.1
opentelemetry-instrumentation-writer==0.50.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b1
packaging==25.0
pandas==2.3.3
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.21
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.1.2
starlette==0.50.0
syncer==2.0.3
tenacity==9.1.2
tokenizers==0.22.2
tomli==2.4.0
tqdm==4.67.1
traceloop-sdk==0.50.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
wrapt==1.17.3
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
